+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './logger';
 export * from './reader';
 export * from './dbError';
+export * from './sqlGuard';